+++ packages/react/src/Overlay/Overlay.tsx
@@ -1,20 +1,17 @@
 import type {ComponentPropsWithRef, ReactElement} from 'react'
 import React, {useEffect, useRef} from 'react'
 import useLayoutEffect from '../utils/useIsomorphicLayoutEffect'
-import {get} from '../constants'
 import type {AriaRole, Merge} from '../utils/types'
 import type {TouchOrMouseEvent} from '../hooks'
 import {useOverlay} from '../hooks'
 import Portal from '../Portal'
-import type {SxProp} from '../sx'
 import {useRefObjectAsForwardedRef} from '../hooks/useRefObjectAsForwardedRef'
 import type {AnchorSide} from '@primer/behaviors'
-import {useTheme} from '../ThemeProvider'
 import type {ForwardRefComponent as PolymorphicForwardRefComponent} from '../utils/polymorphic'
 import {useFeatureFlag} from '../FeatureFlags'
 import classes from './Overlay.module.css'
 import {clsx} from 'clsx'
-import {BoxWithFallback} from '../internal/components/BoxWithFallback'
+import theme from '../theme'
 
 type StyledOverlayProps = {
   width?: keyof typeof widthMap
@@ -24,7 +21,7 @@ type StyledOverlayProps = {
   visibility?: 'visible' | 'hidden'
   overflow?: 'auto' | 'hidden' | 'scroll' | 'visible'
   style?: React.CSSProperties
-} & SxProp
+}
 
 export const heightMap = {
   xsmall: '192px',
@@ -91,7 +88,7 @@ type OwnOverlayProps = Merge<StyledOverlayProps, BaseOverlayProps>
  * @param bottom Optional. Vertical bottom position of the overlay, relative to its closest positioned ancestor (often its `Portal`).
  * @param position Optional. Sets how an element is positioned in a document. Defaults to `absolute` positioning.
  */
-export const BaseOverlay = React.forwardRef<HTMLDivElement, OwnOverlayProps>(
+export const BaseOverlay = React.forwardRef(
   (
     {
       visibility,
@@ -106,13 +103,13 @@ export const BaseOverlay = React.forwardRef<HTMLDivElement, OwnOverlayProps>(
       className,
       maxHeight,
       maxWidth,
+      as: Component = 'div',
       ...rest
     },
     forwardedRef,
   ): ReactElement => {
     return (
-      <BoxWithFallback
-        as="div"
+      <Component
         {...rest}
         ref={forwardedRef}
         style={
@@ -194,9 +191,8 @@ const Overlay = React.forwardRef<HTMLDivElement, internalOverlayProps>(
   ): ReactElement => {
     const overlayRef = useRef<HTMLDivElement>(null)
     useRefObjectAsForwardedRef(forwardedRef, overlayRef)
-    const {theme} = useTheme()
-    const slideAnimationDistance = parseInt(get('space.2')(theme).replace('px', ''))
-    const slideAnimationEasing = get('animation.easeOutCubic')(theme)
+    const slideAnimationDistance = parseInt(theme.space[2], 10)
+    const slideAnimationEasing = theme.animation.easeOutCubic
 
     useOverlay({
       overlayRef,