+++ packages/styled-react/src/__tests__/primer-react-deprecated.browser.test.tsx
@@ -5,12 +5,19 @@ import {Button} from '../index'
 
 describe('@primer/react/deprecated', () => {
   test('Dialog supports `sx` prop', () => {
-    render(<Dialog data-testid="component" isOpen sx={{background: 'red'}} />)
+    render(<Dialog as="button" data-testid="component" isOpen sx={{background: 'red'}} />)
     expect(window.getComputedStyle(screen.getByTestId('component')).backgroundColor).toBe('rgb(255, 0, 0)')
+    expect(screen.getByTestId('component').role).toBe('dialog')
+  })
+
+  test('Dialog.Header supports `sx` prop', () => {
+    render(<Dialog.Header as="button" data-testid="component" sx={{background: 'red'}} />)
+    expect(window.getComputedStyle(screen.getByTestId('component')).backgroundColor).toBe('rgb(255, 0, 0)')
+    expect(screen.getByTestId('component').className.includes('Header')).toBe(true)
   })
 
   test('Octicon supports `sx` prop', () => {
-    render(<Octicon data-testid="component" icon={props => <svg {...props} />} sx={{background: 'red'}} />)
+    render(<Octicon data-testid="component" as="button" icon={props => <svg {...props} />} sx={{background: 'red'}} />)
     expect(window.getComputedStyle(screen.getByTestId('component')).backgroundColor).toBe('rgb(255, 0, 0)')
   })
 
@@ -21,6 +28,7 @@ describe('@primer/react/deprecated', () => {
 
   test('TabNav.Link supports `sx` prop', () => {
     render(<TabNav.Link data-testid="component" sx={{background: 'red'}} as={Button} />)
+    expect(screen.getByTestId('component')).toHaveAttribute('role', 'tab')
     expect(window.getComputedStyle(screen.getByTestId('component')).backgroundColor).toBe('rgb(255, 0, 0)')
     expect(window.getComputedStyle(screen.getByRole('tab')).backgroundColor).toBe('rgb(255, 0, 0)')
     expect(screen.getByRole('tab').tagName).toBe('BUTTON')