+++ packages/styled-react/src/components/deprecated/TabNav.tsx
@@ -1,8 +1,8 @@
 import {TabNav as PrimerTabNav} from '@primer/react/deprecated'
 import type {TabNavProps as PrimerTabNavProps, TabNavLinkProps as PrimerTabNavLinkProps} from '@primer/react/deprecated'
-import {sx, type SxProp} from '../sx'
+import {sx, type SxProp} from '../../sx'
 import styled from 'styled-components'
-import {type ForwardRefComponent} from '../polymorphic'
+import {forwardRef} from 'react'
 
 type TabNavProps = PrimerTabNavProps & SxProp
 type TabNavLinkProps = PrimerTabNavLinkProps & SxProp
@@ -13,17 +13,19 @@ const StyledTabNav = styled(PrimerTabNav).withConfig({
   ${sx}
 `
 
-// @ts-ignore forwardedAs is valid here but I don't know how to fix the typescript error
-const TabNavImpl = ({as, ...props}: TabNavProps) => <StyledTabNav forwardedAs={as} {...props} />
+const TabNavImpl = ({as, ...props}: TabNavProps) => {
+  return <StyledTabNav {...props} {...(as ? {forwardedAs: as as React.ElementType} : {})} />
+}
 
-const StyledTabNavLink: ForwardRefComponent<'a', TabNavLinkProps> = styled(PrimerTabNav.Link).withConfig({
+const StyledTabNavLink = styled(PrimerTabNav.Link).withConfig({
   shouldForwardProp: prop => (prop as keyof TabNavLinkProps) !== 'sx',
 })<TabNavLinkProps>`
   ${sx}
 `
 
-// @ts-ignore forwardedAs is valid here but I don't know how to fix the typescript error
-const TabNavLink = ({as, ...props}: TabNavLinkProps) => <StyledTabNavLink forwardedAs={as} {...props} />
+const TabNavLink = forwardRef<HTMLAnchorElement, TabNavLinkProps>(({as, ...props}, ref) => (
+  <StyledTabNavLink {...props} {...(as ? {forwardedAs: as} : {})} ref={ref} />
+))
 
 const TabNav = Object.assign(TabNavImpl, {
   Link: TabNavLink,