+++ packages/styled-react/src/deprecated.tsx
@@ -1,2 +1,10 @@
-export {TabNav, type TabNavProps, type TabNavLinkProps} from './components/TabNav'
-export {Dialog, Octicon, Tooltip, type TooltipProps} from '@primer/react/deprecated'
+export {TabNav, type TabNavProps, type TabNavLinkProps} from './components/deprecated/TabNav'
+export {Dialog, type DialogProps, type DialogHeaderProps} from './components/deprecated/DialogV1'
+export {Octicon, type OcticonProps} from './components/deprecated/Octicon'
+export {Tooltip, type TooltipProps} from './components/deprecated/Tooltip'
+export {
+  ActionList,
+  type ActionListProps,
+  type ActionListItemProps,
+  type ActionListGroupProps,
+} from './components/deprecated/ActionList'