+++ packages/styled-react/src/experimental.tsx
@@ -5,4 +5,26 @@ export {
   type PageHeaderActionsProps,
   type PageHeaderTitleProps,
 } from './components/PageHeader'
-export {Table, Tooltip, UnderlinePanels} from '@primer/react/experimental'
+
+export {Tooltip, type TooltipProps} from './components/Tooltip'
+
+export {
+  UnderlinePanels,
+  type UnderlinePanelsProps,
+  type UnderlinePanelsTabProps,
+  type UnderlinePanelsPanelProps,
+} from './components/UnderlinePanels'
+
+export {
+  Table,
+  type DataTableContainerProps,
+  type TableProps,
+  type TableHeadProps,
+  type TableBodyProps,
+  type TableRowProps,
+  type TableHeaderProps,
+  type TableCellProps,
+  type TableTitleProps,
+  type TableSubtitleProps,
+  type TableActionsProps,
+} from './components/DataTable'