+++ packages/styled-react/src/index.tsx
@@ -1,44 +1,50 @@
-export {ActionList} from '@primer/react'
-export {ActionMenu} from '@primer/react'
-export {Autocomplete} from '@primer/react'
-export {Avatar} from '@primer/react'
-export {Breadcrumbs} from '@primer/react'
+'use client'
+
 export {Box, type BoxProps} from './components/Box'
 export {Button} from '@primer/react'
-export {CheckboxGroup} from '@primer/react'
-export {CircleBadge} from '@primer/react'
 export {Details} from '@primer/react'
 export {FormControl} from '@primer/react'
-export {Heading} from '@primer/react'
 export {IconButton} from '@primer/react'
-export {Label} from '@primer/react'
-export {Overlay} from '@primer/react'
 export {ProgressBar} from '@primer/react'
 export {Select} from '@primer/react'
-export {Text} from '@primer/react'
 export {Textarea} from '@primer/react'
 export {TextInput} from '@primer/react'
-export {Token} from '@primer/react'
-export {Tooltip} from '@primer/react'
 export {type TextInputProps} from '@primer/react'
-export {type TokenProps} from '@primer/react'
 
 // theming depends on styled-components
-export {ThemeProvider} from '@primer/react'
+export {ThemeProvider, useTheme, useColorSchemeVar, BaseStyles} from './components/FeatureFlaggedTheming'
 export {merge} from '@primer/react'
 export {theme} from '@primer/react'
 export {themeGet} from '@primer/react'
-export {useColorSchemeVar} from '@primer/react'
-export {useTheme} from '@primer/react'
 
+export {
+  ActionList,
+  type ActionListProps,
+  type ActionListItemProps,
+  type ActionListLinkItemProps,
+  type ActionListGroupProps,
+  type ActionListDividerProps,
+  type ActionListLeadingVisualProps,
+  type ActionListTrailingVisualProps,
+  type ActionListTrailingActionProps,
+} from './components/ActionList'
+export {ActionMenu} from './components/ActionMenu'
+export {Autocomplete, type AutocompleteOverlayProps} from './components/Autocomplete'
+export {Avatar, type AvatarProps} from './components/Avatar'
+export {Breadcrumbs, Breadcrumb, type BreadcrumbsProps, type BreadcrumbsItemProps} from './components/Breadcrumbs'
 export {Checkbox, type CheckboxProps} from './components/Checkbox'
+export {CheckboxGroup, type CheckboxGroupProps} from './components/CheckboxGroup'
+export {CircleBadge} from './components/CircleBadge'
 export {CounterLabel, type CounterLabelProps} from './components/CounterLabel'
 export {Dialog, type DialogProps} from './components/Dialog'
 export {Flash} from './components/Flash'
 export {Header, type HeaderProps} from './components/Header'
+export {Heading} from './components/Heading'
+export {Label, type LabelProps} from './components/Label'
 export {Link, type LinkProps} from './components/Link'
 export {LinkButton, type LinkButtonProps} from './components/LinkButton'
 export {NavList, type NavListProps} from './components/NavList'
+export {Overlay} from './components/Overlay'
 export {PageLayout, type PageLayoutProps} from './components/PageLayout'
 export {
   PageHeader,
@@ -57,6 +63,7 @@ export {
 export {Spinner, type SpinnerProps} from './components/Spinner'
 export {StateLabel, type StateLabelProps} from './components/StateLabel'
 export {SubNav, type SubNavProps, type SubNavLinkProps} from './components/SubNav'
+export {Text, type TextProps} from './components/Text'
 export {
   Timeline,
   type TimelineProps,
@@ -66,6 +73,8 @@ export {
   type TimelineBreakProps,
 } from './components/Timeline'
 export {ToggleSwitch, type ToggleSwitchProps} from './components/ToggleSwitch'
+export {Tooltip, type TooltipProps} from './components/Tooltip'
+export {Token, type TokenProps} from './components/Token'
 export {Truncate, type TruncateProps} from './components/Truncate'
 export {UnderlineNav, type UnderlineNavProps, type UnderlineNavItemProps} from './components/UnderlineNav'